+++ src/main/java/insane96mcp/survivalreimagined/data/lootmodifier/SRGlobalLootModifierProvider.java
@@ -5,6 +5,7 @@ import insane96mcp.survivalreimagined.module.farming.feature.Crops;
 import insane96mcp.survivalreimagined.module.farming.feature.Livestock;
 import insane96mcp.survivalreimagined.module.items.feature.FlintTools;
 import insane96mcp.survivalreimagined.module.mining.feature.Iron;
+import insane96mcp.survivalreimagined.module.mobs.feature.Equipment;
 import insane96mcp.survivalreimagined.module.world.feature.Loot;
 import net.minecraft.data.PackOutput;
 import net.minecraftforge.common.data.GlobalLootModifierProvider;
@@ -20,6 +21,7 @@ public class SRGlobalLootModifierProvider extends GlobalLootModifierProvider {
         Crops.addGlobalLoot(this);
         EnchantmentsFeature.addGlobalLoot(this);
         Livestock.addGlobalLoot(this);
+        Equipment.addGlobalLoot(this);
         Loot.addGlobalLoot(this);
         FlintTools.addGlobalLoot(this);
     }