+++ src/main/java/insane96mcp/survivalreimagined/module/experience/feature/EnchantmentsFeature.java
@@ -54,9 +54,6 @@ public class EnchantmentsFeature extends Feature {
 	@Label(name = "Infinity overhaul", description = "Infinity can go up to level 4. Each level makes an arrow have 1 in level chance to not consume.")
 	public static Boolean infinityOverhaul = true;
 	@Config
-	@Label(name = "Unbreaking overhaul", description = "Unbreaking max level is set to 1.")
-	public static Boolean unbreakingOverhaul = true;
-	@Config
 	@Label(name = "Replace Bane of Arthropods", description = "Replace Bane of Arthropods with Bane of SSSSS, similar to Bane of Arthropods deals more damage to Spiders but also creepers.")
 	public static Boolean replaceBaneOfArthropods = true;
 
@@ -215,10 +212,6 @@ public class EnchantmentsFeature extends Feature {
 
 	public static boolean isInfinityOverhaulEnabled() {
 		return Feature.isEnabled(EnchantmentsFeature.class) && infinityOverhaul;
-	}
-
-	public static boolean isUnbreakingOverhaulEnabled() {
-		return Feature.isEnabled(EnchantmentsFeature.class) && unbreakingOverhaul;
 	}
 
 	private static final String path = "enchantments/";