+++ changelog.md
@@ -1,6 +1,8 @@
 # Changelog
 
 ## Upcoming
+* Breaking a tree without an axe now drops planks
+* Buckets now require the Nether to be made
 * Removed Gold Nuggets drop from Zombified Piglins
 
 ## Alpha 3.5.0